requirements.txt
flask>=3.1
python-dotenv>=1.1
filelock>=3.16
gunicorn>=23.0
waitress>=3.0
pyodbc>=5.1
boto3>=1.35
requests>=2.31
